console: Refactor connect in model select
